New Analysis Method: Transient
Transisent creates plots of Channel vs Time above Threshold. Ie time
with Engine RPM > 1000. This allows plots to be created for
investigating the transient response of the various parameters

The resulting plot looks like a bivariate histogram, but normalization
only occurs in each time slice. As Transisent seeks to show the steady
state of the channel data and not the most freqent time + value
combination

Example and documentation will be coming soonish